+++ src/components/DatePicker/DatePicker/DatePicker.vue
@@ -4,7 +4,7 @@
 	import DateTextInput from '../DateTextInput/DateTextInput.vue'
 	import ComplexDatePicker from '../ComplexDatePicker/ComplexDatePicker.vue'
 	import { VDatePicker } from 'vuetify/components'
-	import { vAccessibleButtons } from '../accessibleButtons'
+	import { vAccessibleButtons, vCleanAriaAttributes } from '../directives'
 	import { useValidation } from '@/composables/validation/useValidation'
 	import { useDateFormat } from '@/composables/date/useDateFormatDayjs'
 	import { useDateInitialization, type DateValue, type DateInput } from '@/composables/date/useDateInitializationDayjs'
@@ -854,6 +854,7 @@
 
 <template>
 	<div
+		v-clean-aria-attributes
 		class="date-picker-container"
 		:style="inputStyle"
 	>
@@ -980,9 +981,9 @@
 				</template>
 				<VDatePicker
 					v-if="isDatePickerVisible && !props.noCalendar"
-					v-accessible-buttons
 					ref="datePickerRef"
 					v-model="selectedDates"
+					v-accessible-buttons
 					color="primary"
 					:first-day-of-week="1"
 					:multiple="props.displayRange ? 'range' : false"
