+++ src/components/DatePicker/directives/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Export des directives d'accessibilité pour le DatePicker
+ */
+
+export { vAccessibleButtons } from './accessibleButtons'
+export { vCleanAriaAttributes } from './cleanAriaAttributes'
